메뉴 아이콘 추가 및 링크 연결, css 수정

+++ src/main/java/com/itwillbs/factron/repository/vacation/VacationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VacationRepository extends JpaRepository<VacationHistory, Integer> {
+public interface VacationRepository extends JpaRepository<VacationHistory, Long> {
 }
